+++ apps/core/src/plugins/oauth2/utils/objectKeys.ts
